requirements.txt
alembic==1.13.1
blinker==1.8.2
blueprint==3.4.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pycparser==2.22
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.0
Werkzeug==3.0.3
WTForms==3.1.2